ref: Remove timelines/ folder and rename main.js as index.js
