+++ src/App/components/JsPsychExperiment.jsx
@@ -2,8 +2,8 @@ import { initJsPsych } from "jspsych";
 import React from "react";
 
 import { config, taskVersion } from "../../config/main";
+import { buildTimeline, jsPsychOptions } from "../../experiment";
 import { initParticipant } from "../deployments/firebase";
-import { buildTimeline, jsPsychOptions } from "../../experiment/timelines/main";
 
 // ID used to identify the DOM element that holds the experiment.
 const EXPERIMENT_ID = "experiment-window";
